Improved build time 10x
With Swift 4 the complexity (from the point of view of the type checker)
of an expression with mixed int sizes grows pretty fast when we start
adding additional sub-expressions.

This update removes a lot of castings to u32/u64 for the Int and UInt
extensions reducing by 90% the complexity of those methods.

The long expressions have not been broken down to separated
sub-expressions since the gains seems to be minimal at the moment.